requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
docopt==0.6.2
future==0.18.2
idna==3.3
Markdown==3.3.7
pipreqs==0.4.11
pycodestyle==2.8.0
pytz==2021.3
requests==2.27.1
serializers==0.2.4
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
yarg==0.1.9
